array: split up functionality into free functions
One of the goals I have for this project is to perform well in the
absence of heavy optimization. One of the ways I want to do this is to
try to minimize function call overhead for things which are called
extremely often. What I've done here is taken some common functionality
from the various XArray types and consolidated them into #[inline] free
functions. This means that each time these methods are called, even
with no optimization, there shouldn't be more than one level of
function calls before the actual logic is called. We don't need to rely
on optimization to do this, and we know we want it anyway since all of
these functions are small.